Restructured sync script to prepare for a proper architecture
