+++ scripts/sync/utils/logger.ts
