+++ scripts/sync/utils/minificator.ts
